patch torch_chunk_gated_delta_rule
