+++ vllm_ascend/patch/worker/patch_common/patch_triton.py
@@ -1,10 +1,11 @@
+import vllm.model_executor.layers.fla.ops.chunk
 import vllm.model_executor.layers.fla.ops.fused_recurrent
 import vllm.model_executor.layers.fla.ops.layernorm_guard
 import vllm.model_executor.layers.mamba.ops.causal_conv1d
 
 from vllm_ascend.ops.casual_conv1d import (causal_conv1d_fn,
                                            causal_conv1d_update_npu)
-from vllm_ascend.ops.fla import LayerNormFn
+from vllm_ascend.ops.fla import LayerNormFn, torch_chunk_gated_delta_rule
 from vllm_ascend.ops.sigmoid_gating import \
     fused_recurrent_gated_delta_rule_fwd_kernel
 
@@ -12,3 +13,4 @@ vllm.model_executor.layers.mamba.ops.causal_conv1d.causal_conv1d_update = causal
 vllm.model_executor.layers.mamba.ops.causal_conv1d.causal_conv1d_fn = causal_conv1d_fn
 vllm.model_executor.layers.fla.ops.fused_recurrent.fused_recurrent_gated_delta_rule_fwd_kernel = fused_recurrent_gated_delta_rule_fwd_kernel
 vllm.model_executor.layers.fla.ops.layernorm_guard.LayerNormFn = LayerNormFn
+vllm.model_executor.layers.fla.ops.chunk.chunk_gated_delta_rule = torch_chunk_gated_delta_rule
